+++ src/com/epam/exception/NotExistEntityException.java
@@ -1,7 +1,7 @@
 package com.epam.exception;
 
 public class NotExistEntityException extends PeriodicalException {
-    public NotExistEntityException(int id) {
+    public NotExistEntityException(Long id) {
         super("THE ENTITY WITH ID = " + id + " ARE NOT EXIST !");
     }
 }
