+++ src/com/epam/model/periodical/Publisher.java
@@ -4,10 +4,10 @@ import java.io.Serializable;
 import java.util.Objects;
 
 public class Publisher implements Serializable {
-    private int id;
+    private Long id;
     private String name;
 
-    public Publisher(int id, String name) {
+    public Publisher(Long id, String name) {
         this.id = id;
         this.name = name;
     }
@@ -16,11 +16,11 @@ public class Publisher implements Serializable {
         this.name = name;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
